requirements.txt
dspy-ai==2.4.3
jinja2==3.1.3
structlog==24.1.0
chromadb==0.4.24
langchain==0.1.15
